Добавляет модульность и чинит ошибки из раздела 4
